cut a lot of things
